refactor(database): :recycle: separation queries.py to different files
Файл queries.py был разделён на несколько других файлов .py по смыслу функций
