+++ database/stats.py
@@ -4,6 +4,15 @@ from json import dumps, loads
 from database.core import db
 
 
+async def format_currency(val: float) -> str:
+    return f"{val:,.2f}".replace(",", " ")
+
+
+from aiogram.types import CallbackQuery
+
+from keyboards.inline import items_buttons
+
+
 class StatsManager:
     """
     Класс для работы со статистикой.
@@ -32,29 +41,38 @@ class StatsManager:
 
     @staticmethod
     def _get_today(*, is_additional: bool = False) -> str:
-        now = datetime.datetime.now(tz=datetime.UTC)
+        now = datetime.datetime.now()
         if is_additional:
-            return now.isoformat(timespec='seconds').split('+')[0]
+            return now.isoformat(timespec='seconds')
         else:
             return now.date().isoformat()
 
     def _init_currency_fields(self, currency: str):
         keys = ["invested", "sold", "average_buy_price", "average_sell_price", "roi", "profit"]
         for key in keys:
-            self.stats.setdefault(key, {})[currency] = self.stats.setdefault(key, {}).get(currency, 0)
+            if key not in self.stats:
+                self.stats[key] = {}
+            if currency not in self.stats[key]:
+                self.stats[key][currency] = 0
 
-        self.stats.setdefault("profit", {})["total"] = self.stats.setdefault("profit", {}).get("total", 0)
-
-        self.stats.setdefault("current_portfolio", {})[currency] = self.stats.setdefault("current_portfolio", {}).get(currency, {
-            "amount": 0,
-            "avg_price": 0
-        })
+        if "profit" not in self.stats:
+            self.stats["profit"] = {}
+        if "total" not in self.stats["profit"]:
+            self.stats["profit"]["total"] = 0
+        if "current_portfolio" not in self.stats:
+            self.stats["current_portfolio"] = {}
+        if currency not in self.stats["current_portfolio"]:
+            self.stats["current_portfolio"][currency] = {"amount": 0, "avg_price": 0}
+        
+        if "sell_history" not in self.stats:
+            self.stats["sell_history"] = {}
+        if currency not in self.stats["sell_history"]:
+            self.stats["sell_history"][currency] = {"total_amount": 0, "total_value": 0}
 
     def _init_counters(self):
         self.stats.setdefault("deal_count", {"total": 0, "buy": 0, "sell": 0})
         self.stats.setdefault("spam_attempts", 0)
         self.stats.setdefault("profit_by_day", {})
-        self.stats.setdefault("favorite_currency", None)
 
     async def update_favorite_currency(self):
         totals = {}
@@ -67,6 +85,9 @@ class StatsManager:
             self.stats["favorite_currency"] = max(totals, key=totals.get)
 
     async def record_buy(self, currency: str, amount: float, price: float):
+        if amount <= 0 or price <= 0:
+            raise ValueError("Количество и цена должны быть положительными числами")
+        
         self._init_counters()
         self._init_currency_fields(currency)
 
@@ -77,7 +98,7 @@ class StatsManager:
         current = self.stats["current_portfolio"].get(currency, {"amount": 0, "avg_price": 0})
         total_amount = current["amount"] + amount
         total_value = current["amount"] * current["avg_price"] + amount * price
-        avg_price = total_value / total_amount if total_amount != 0 else price
+        avg_price = total_value / total_amount if total_amount > 0 else 0
 
         self.stats["current_portfolio"][currency] = {
             "amount": round(total_amount, 4),
@@ -100,14 +121,19 @@ class StatsManager:
         self._init_counters()
         self._init_currency_fields(currency)
 
-        lots = self.trades[currency]
+        lots = self.trades[currency].copy()
         new_lots = []
         profit = 0.0
         invested_in_sold = 0.0
 
         for lot in lots:
+            if remaining <= 0:
+                new_lots.append(lot)
+                continue
+                
             lot_amount = lot["amount"]
             lot_price = lot["price"]
+            
             if lot_amount <= remaining:
                 profit += lot_amount * (price - lot_price)
                 invested_in_sold += lot_amount * lot_price
@@ -115,10 +141,10 @@ class StatsManager:
             else:
                 profit += remaining * (price - lot_price)
                 invested_in_sold += remaining * lot_price
-                lot["amount"] -= remaining
-                new_lots.append(lot)
+                updated_lot = lot.copy()
+                updated_lot["amount"] = lot_amount - remaining
+                new_lots.append(updated_lot)
                 remaining = 0
-                break
 
         if remaining > 0:
             raise ValueError("Недостаточно валюты для продажи")
@@ -126,6 +152,9 @@ class StatsManager:
         return profit, invested_in_sold, new_lots
 
     async def record_sell(self, currency: str, amount: float, price: float):
+        if amount <= 0 or price <= 0:
+            raise ValueError("Количество и цена должны быть положительными числами")
+        
         self._init_counters()
         self._init_currency_fields(currency)
 
@@ -140,7 +169,6 @@ class StatsManager:
 
         self.trades[currency] = new_lots
 
-        # Обновляем текущий портфель
         current = self.stats["current_portfolio"].get(currency, {"amount": 0, "avg_price": 0})
         total_amount = current["amount"] - amount
         avg_price = current["avg_price"] if total_amount > 0 else 0
@@ -150,7 +178,21 @@ class StatsManager:
             "avg_price": round(avg_price, 2)
         }
 
-        self.stats["average_sell_price"][currency] = round(price, 2)
+        sell_history = self.stats["sell_history"][currency]
+        old_total_amount = sell_history["total_amount"]
+        old_total_value = sell_history["total_value"]
+        
+        new_total_amount = old_total_amount + amount
+        new_total_value = old_total_value + (amount * price)
+        
+        self.stats["sell_history"][currency] = {
+            "total_amount": new_total_amount,
+            "total_value": new_total_value
+        }
+        
+        self.stats["average_sell_price"][currency] = round(
+            new_total_value / new_total_amount if new_total_amount > 0 else 0, 2
+        )
 
         self.stats["profit"][currency] += profit
         self.stats["profit"]["total"] += profit
@@ -159,20 +201,23 @@ class StatsManager:
         self.stats["profit_by_day"].setdefault(today, 0)
         self.stats["profit_by_day"][today] += profit
 
-        roi = (profit / invested_in_sold) * 100 if invested_in_sold != 0 else 0
+        total_invested = self.stats["invested"].get(currency, 0)
+        total_profit = self.stats["profit"].get(currency, 0)
+        roi = (total_profit / total_invested) * 100 if total_invested > 0 else 0
         self.stats["roi"][currency] = round(roi, 2)
 
         current_best = self.stats.get("best_deal", None)
+        deal_roi = (profit / invested_in_sold) * 100 if invested_in_sold > 0 else 0
         deal_info = {
             "currency": currency,
             "amount": amount,
             "price": price,
             "profit": profit,
-            "roi": round(roi, 2),
+            "roi": round(deal_roi, 2),
             "date": self._get_today(is_additional=False)
         }
-        # Если еще нет лучшей сделки или новая прибыль выше текущей
-        if (current_best is None) or (deal_info["roi"] > current_best.get("roi", float("-inf"))):
+        
+        if (current_best is None) or (profit > current_best.get("profit", 0)):
             self.stats["best_deal"] = deal_info
 
         await self.update_favorite_currency()
@@ -186,3 +231,46 @@ class StatsManager:
 
     async def save(self):
         await self.db.update_data("users", {"stats": dumps(self.stats), "trades": dumps(self.trades)}, {"id": self.user_id})
+
+    async def show_stats(self, username: str, user_id: int, call: CallbackQuery) -> None:
+        stats = StatsManager(user_id)
+        await stats.load()
+
+        deal_count = stats.stats.get('deal_count', {})
+        invested = stats.stats.get('invested', {})
+        sold = stats.stats.get('sold', {})
+        profit = stats.stats.get('profit', {})
+        roi = stats.stats.get('roi', {})
+        best_deal = stats.stats.get('best_deal', {})
+
+        text = (
+            f"📊\u003cb\u003eСтатистика пользователя @{username}\u003c/b\u003e (\u003ci\u003e{user_id}\u003c/i\u003e)\n\n"
+            "🏛 \u003cu\u003eОперации:\u003c/u\u003e\n"
+            f"• Покупка: {deal_count.get('buy', 0)}\n"
+            f"• Продажа: {deal_count.get('sell', 0)}\n"
+            f"• Общее: {deal_count.get('total', 0)}\n\n"
+            "💸 \u003cu\u003eИнвестировано:\u003c/u\u003e\n"
+            f"• ST: {await format_currency(invested.get('st', 0))} RUB\n"
+            f"• V: {await format_currency(invested.get('v', 0))} RUB\n\n"
+            "💰 \u003cu\u003eПродано:\u003c/u\u003e\n"
+            f"• ST: {await format_currency(sold.get('st', 0))} RUB\n"
+            f"• V: {await format_currency(sold.get('v', 0))} RUB\n\n"
+            "📈 \u003cu\u003eПрибыль:\u003c/u\u003e\n"
+            f"• ST: {await format_currency(profit.get('st', 0))} RUB\n"
+            f"• V: {await format_currency(profit.get('v', 0))} RUB\n\n"
+            "📊 \u003cu\u003eROI*:\u003c/u\u003e\n"
+            f"• ST: {roi.get('st', 0)}%\n"
+            f"• V: {roi.get('v', 0)}%\n\n"
+        )
+
+        if stats.stats.get('favorite_currency'):
+            text += f"⭐️ \u003cu\u003eЛюбимая валюта:\u003c/u\u003e {stats.stats.get('favorite_currency', '').upper()}\n"
+        if best_deal:
+            text += f"🏆 \u003cu\u003eЛучшая сделка:\u003c/u\u003e {best_deal.get('currency', '').upper()} \u003ci\u003e({best_deal.get('roi', 0)}% ROI*)\u003c/i\u003e\n"
+
+        text += (
+            f"😡 \u003cu\u003eПопытки спама:\u003c/u\u003e {stats.stats.get('spam_attempts', 0)}\n\n"
+            "\u003ci\u003e*ROI (Return on Investment) — доходность инвестиций в процентах\u003c/i\u003e"
+        )
+
+        await call.message.edit_text(text, reply_markup=items_buttons)
